backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.3.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2.7
django-cors-headers==4.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
joblib==1.3.2
kombu==5.6.2
numpy==1.26.3
packaging==26.0
pandas==2.1.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
scikit-learn==1.3.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
vine==5.1.0
wcwidth==0.6.0
